requirements.txt
fastapi==0.115.4
uvicorn[standard]
pydantic
scrapy
beautifulsoup4
lxml
dask[distributed]
confluent-kafka
pymongo[srv]
qdrant-client
langchain
tenacity
python-dotenv
orjson
requests
faiss-cpu
ray
pika
numpy
langchain-google-genai
